# api/openapi.yaml
openapi: 3.1.0
info:
  title: WorkBuddy Expert Team OS
  version: 1.0.0
paths:
...
  /v1/pilot-programs:
    get:
      tags:
      - production-pilot
      summary: List Programs
      operationId: list_programs_v1_pilot_programs_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
    post:
      tags:
      - production-pilot
      summary: Create Program Route
      operationId: create_program_route_v1_pilot_programs_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PilotProgramIn'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/pilot-programs/{program_id}/transition:
    post:
      tags:
      - production-pilot
      summary: Transition Program Route
      operationId: transition_program_route_v1_pilot_programs__program_id__transition_post
      parameters:
      - name: program_id
        in: path
        required: true
        schema:
          type: string
          title: Program Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/workbuddy__api__pilot_routes__TransitionIn'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/pilot-programs/{program_id}:
    get:
      tags:
      - production-pilot
      summary: Get Program
      operationId: get_program_v1_pilot_programs__program_id__get
      parameters:
      - name: program_id
        in: path
        required: true
        schema:
          type: string
          title: Program Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/pilot-programs/{program_id}/mailboxes:
    post:
      tags:
      - production-pilot
      summary: Register Mailbox Route
      operationId: register_mailbox_route_v1_pilot_programs__program_id__mailboxes_post
      parameters:
      - name: program_id
        in: path
        required: true
        schema:
          type: string
          title: Program Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PilotMailboxIn'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/pilot-programs/{program_id}/evidence:
    get:
      tags:
      - production-pilot
      summary: List Evidence
      operationId: list_evidence_v1_pilot_programs__program_id__evidence_get
      parameters:
      - name: program_id
        in: path
        required: true
        schema:
          type: string
          title: Program Id
      - name: gate
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Gate
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - production-pilot
      summary: Submit Evidence Route
      operationId: submit_evidence_route_v1_pilot_programs__program_id__evidence_post
      parameters:
      - name: program_id
        in: path
        required: true
        schema:
          type: string
          title: Program Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EvidenceIn'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/pilot-programs/{program_id}/evidence/upload:
    post:
      tags:
      - production-pilot
      summary: Upload Evidence Artifact
      operationId: upload_evidence_artifact_v1_pilot_programs__program_id__evidence_upload_post
      parameters:
      - name: program_id
        in: path
        required: true
        schema:
          type: string
          title: Program Id
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Body_upload_evidence_artifact_v1_pilot_programs__program_id__evidence_upload_post'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/pilot-programs/evidence/{evidence_id}/decision:
    post:
      tags:
      - production-pilot
      summary: Verify Evidence Route
      operationId: verify_evidence_route_v1_pilot_programs_evidence__evidence_id__decision_post
      parameters:
      - name: evidence_id
        in: path
        required: true
        schema:
          type: string
          title: Evidence Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/workbuddy__api__pilot_routes__EvidenceDecisionIn'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /v1/pilot-programs/{program_id}/drills:
    post:
      tags:
      - production-pilot
      summary: Create Drill Route
      operationId: create_drill_route_v1_pilot_programs__program_id__drills_post
      parameters:
      - name: program_id
        in: path
        required: true
        schema:
          type: string
          title: Program Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DrillIn'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    get:
      tags:
      - production-pilot
      summary: List Drills
      operationId: list_drills_v1_pilot_programs__program_id__drills_get
      parameters:
      - name: program_id
        in: path
        required: true
        schema:
          type: string
          title: Program Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/pilot-programs/drills/{drill_id}/complete:
    post:
      tags:
      - production-pilot
      summary: Complete Drill Route
      operationId: complete_drill_route_v1_pilot_programs_drills__drill_id__complete_post
      parameters:
      - name: drill_id
        in: path
        required: true
        schema:
          type: string
          title: Drill Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DrillCompleteIn'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/pilot-programs/{program_id}/attestations:
    post:
      tags:
      - production-pilot
      summary: Attest Route
      operationId: attest_route_v1_pilot_programs__program_id__attestations_post
      parameters:
      - name: program_id
        in: path
        required: true
        schema:
          type: string
          title: Program Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AttestationIn'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    get:
      tags:
      - production-pilot
      summary: List Attestations
      operationId: list_attestations_v1_pilot_programs__program_id__attestations_get
      parameters:
      - name: program_id
        in: path
        required: true
        schema:
          type: string
          title: Program Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/pilot-programs/{program_id}/gates/{gate_key}:
    get:
      tags:
      - production-pilot
      summary: Gate Status
      operationId: gate_status_v1_pilot_programs__program_id__gates__gate_key__get
      parameters:
      - name: program_id
        in: path
        required: true
        schema:
          type: string
          title: Program Id
      - name: gate_key
        in: path
        required: true
        schema:
          type: string
          title: Gate Key
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/pilot-programs/{program_id}/go-no-go:
    get:
      tags:
      - production-pilot
      summary: Report
      operationId: report_v1_pilot_programs__program_id__go_no_go_get
      parameters:
      - name: program_id
        in: path
        required: true
        schema:
          type: string
          title: Program Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/pilot-programs/{program_id}/incidents:
    post:
      tags:
      - production-pilot
      summary: Incident Route
      operationId: incident_route_v1_pilot_programs__program_id__incidents_post
      parameters:
      - name: program_id
        in: path
        required: true
        schema:
          type: string
          title: Program Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IncidentIn'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/pilot-programs/incidents/{incident_id}/resolve:
    post:
      tags:
      - production-pilot
      summary: Resolve Incident Route
      operationId: resolve_incident_route_v1_pilot_programs_incidents__incident_id__resolve_post
      parameters:
      - name: incident_id
        in: path
        required: true
        schema:
          type: string
          title: Incident Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IncidentResolveIn'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    AttestationIn:
      properties:
        gate_key:
          type: string
          title: Gate Key
        role:
          type: string
          title: Role
        decision:
          type: string
          title: Decision
        notes:
          type: string
          title: Notes
          default: ''
      type: object
      required:
      - gate_key
      - role
      - decision
      title: AttestationIn
    Body_upload_evidence_artifact_v1_pilot_programs__program_id__evidence_upload_post:
      properties:
        gate_key:
          type: string
          title: Gate Key
        evidence_type:
          type: string
          title: Evidence Type
        source:
          type: string
          title: Source
          default: operator
        environment:
          type: string
          title: Environment
          default: production
        file:
          type: string
          contentMediaType: application/octet-stream
          title: File
      type: object
      required:
      - gate_key
      - evidence_type
      - file
      title: Body_upload_evidence_artifact_v1_pilot_programs__program_id__evidence_upload_post
...
    DrillCompleteIn:
      properties:
        passed:
          type: boolean
          title: Passed
        result:
          additionalProperties: true
          type: object
          title: Result
        evidence_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Evidence Id
      type: object
      required:
      - passed
      title: DrillCompleteIn
    DrillIn:
      properties:
        drill_type:
          type: string
          title: Drill Type
        execution_mode:
          type: string
          title: Execution Mode
          default: SIMULATED
      type: object
      required:
      - drill_type
      title: DrillIn
...
    EvidenceIn:
      properties:
        gate_key:
          type: string
          title: Gate Key
        evidence_type:
          type: string
          title: Evidence Type
        source:
          type: string
          title: Source
        environment:
          type: string
          title: Environment
          default: staging
        metrics:
          additionalProperties: true
          type: object
          title: Metrics
        artifact_ref:
          anyOf:
          - type: string
          - type: 'null'
          title: Artifact Ref
      type: object
      required:
      - gate_key
      - evidence_type
      - source
      title: EvidenceIn
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    IncidentIn:
      properties:
        severity:
          type: string
          title: Severity
        category:
          type: string
          title: Category
        title:
          type: string
          title: Title
        details:
          additionalProperties: true
          type: object
          title: Details
      type: object
      required:
      - severity
      - category
      - title
      title: IncidentIn
    IncidentResolveIn:
      properties:
        resolution:
          type: string
          minLength: 3
          title: Resolution
      type: object
      required:
      - resolution
      title: IncidentResolveIn
...
    PilotMailboxIn:
      properties:
        mail_account_id:
          type: string
          title: Mail Account Id
        mode:
          type: string
          title: Mode
          default: SHADOW
        team_keys:
          items:
            type: string
          type: array
          title: Team Keys
        allowed_recipient_domains:
          items:
            type: string
          type: array
          title: Allowed Recipient Domains
        allowed_recipient_addresses:
          items:
            type: string
          type: array
          title: Allowed Recipient Addresses
        daily_send_limit:
          type: integer
          maximum: 1000.0
          minimum: 0.0
          title: Daily Send Limit
          default: 0
      type: object
      required:
      - mail_account_id
      title: PilotMailboxIn
    PilotProgramIn:
      properties:
        name:
          type: string
          maxLength: 240
          minLength: 3
          title: Name
        start_date:
          anyOf:
          - type: string
          - type: 'null'
          title: Start Date
        end_date:
          anyOf:
          - type: string
          - type: 'null'
          title: End Date
        scope:
          additionalProperties: true
          type: object
          title: Scope
        targets:
          additionalProperties: true
          type: object
          title: Targets
        owners:
          additionalProperties:
            anyOf:
            - type: string
            - type: 'null'
          type: object
          title: Owners
      type: object
      required:
      - name
      title: PilotProgramIn
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
    workbuddy__api__pilot_routes__EvidenceDecisionIn:
      properties:
        decision:
          type: string
          title: Decision
        reason:
          type: string
          title: Reason
          default: ''
      type: object
      required:
      - decision
      title: EvidenceDecisionIn
    workbuddy__api__pilot_routes__TransitionIn:
      properties:
        target:
          type: string
          title: Target
      type: object
      required:
      - target
      title: TransitionIn